refactor(core): give bounded response reading one owner
WebFetch bounded its response body with a private helper. The models.dev
refresh needed the same bound and grew a second one that counted
`String.length` — a UTF-16 unit count, which lets a multi-byte body pass a
limit it exceeds by up to three times, and only after the whole body was
already buffered.

Lift WebFetch's implementation to `@maka/core/bounded-response` and have both
call it. The bound stays enforced twice, before and during the read, in the
one place either caller can change it.

Generated-by: Claude Code
